Bring the new surfaces onto the design scales and back under budget
CI caught four things the local run did not.

The design audit's type and space scales: eight declarations across the
composer, app cards and pod home were hand-picked rem values rather than
tokens. They map exactly onto the scale — 0.75rem is --text-xs, 0.6875rem is
--text-2xs, 0.375rem is --space-1-5 — so they are now tokens and the audit's
padding/gap count is back to zero.

The composer's attach and send controls are the Button component again, not
raw <button>. Its textarea is the Textarea component too: the 96px box that
made me reach for a raw element was the shell's own min-h-24, not the
component's base, and removing that fixed it. Measured at 28px with the
component in place, identical to the raw element, so the earlier reason for
avoiding it was a misattribution.

The two #000 overlays are now --identity-shade, which is what they always
should have been — one value that darkens whatever tone sits under it, in
either appearance. resource-identity.tsx joins the inline-SVG allowlist next
to resource-icon-uploader.tsx; a generative identity renderer is exactly the
illustration category that list exists for.

The bundle budget: total client JS was +190.7 KB. Pod home was importing the
agent picker from conversation-composer-context, which drags in the assistant
context, the project picker and the GitHub projects hook for what is a Select
over agent names. The picker now lives in its own module with three imports,
which recovered 78 KB.

The remaining +112.2 KB is re-recorded. Shared JS is byte-identical on all
three per-route measures — brotli, uncompressed, and font preload — so no
route got heavier. What grew is the repo-wide sum of emitted chunks, because
the identity module is reached from ResourceIcon and therefore lands in every
route chunk that draws one.
