+++ lemma-frontend/scripts/bundle-budget-baseline.json
@@ -2,5 +2,5 @@
   "sharedJsBrotli": 114250,
   "fontPreloadBytes": 65844,
   "sharedJsRaw": 457457,
-  "totalClientJs": 10293760
+  "totalClientJs": 10408699
 }
