+++ lemma-frontend/styles/features/composer.css
@@ -112,7 +112,7 @@
     color: var(--text-primary);
 }
 
-.lm-composer-attach {
+.lm-composer-attach.lm-composer-attach {
     display: inline-flex;
     height: 2rem;
     width: 2rem;
@@ -126,7 +126,7 @@
         color var(--dur-control) var(--ease-standard);
 }
 
-.lm-composer-attach:hover:not(:disabled) {
+.lm-composer-attach.lm-composer-attach:hover:not(:disabled) {
     background: var(--bg-subtle);
     color: var(--text-primary);
 }
@@ -136,7 +136,7 @@
     opacity: 0.5;
 }
 
-.lm-composer-send {
+.lm-composer-send.lm-composer-send {
     display: inline-flex;
     height: 2rem;
     width: 2rem;
@@ -151,17 +151,17 @@
         color var(--dur-control) var(--ease-standard);
 }
 
-.lm-composer-send:hover:not(:disabled) {
+.lm-composer-send.lm-composer-send:hover:not(:disabled) {
     background: var(--action-primary-hover);
 }
 
 /* Idle is a filled, muted target — not an absent one. */
-.lm-composer-send[data-state='idle'] {
+.lm-composer-send.lm-composer-send[data-state='idle'] {
     background: var(--surface-3);
     color: var(--text-tertiary);
 }
 
-.lm-composer-send[data-state='busy'] {
+.lm-composer-send.lm-composer-send[data-state='busy'] {
     background: var(--state-error);
     color: var(--text-inverse);
 }
@@ -214,11 +214,11 @@
     min-width: 0;
     height: 1.75rem;
     align-items: center;
-    gap: 0.375rem;
+    gap: var(--space-1-5);
     border: 1px solid var(--border-subtle);
     border-radius: var(--radius-md);
     background: var(--surface-2);
-    padding: 0 0.25rem 0 0.4375rem;
+    padding: 0 var(--space-1) 0 var(--space-1-5);
 }
 
 .lm-file-chip[data-status='failed'] {
@@ -246,7 +246,7 @@
     overflow: hidden;
     min-width: 0;
     color: var(--text-primary);
-    font-size: 0.75rem;
+    font-size: var(--text-xs);
     text-overflow: ellipsis;
     white-space: nowrap;
 }
@@ -254,7 +254,7 @@
 .lm-file-chip-meta {
     flex-shrink: 0;
     color: var(--text-tertiary);
-    font-size: 0.6875rem;
+    font-size: var(--text-2xs);
 }
 
 .lm-file-chip-remove {
