+++ lemma-frontend/styles/features/pod-home.css
@@ -607,7 +607,7 @@
   display: flex;
   min-width: 0;
   flex-direction: column;
-  gap: 0.5rem;
+  gap: var(--space-2);
   border-radius: var(--radius-lg);
   text-decoration: none;
 }
