+++ lemma-frontend/styles/features/resource-ledgers.css
@@ -349,11 +349,11 @@
   z-index: 1;
   display: inline-flex;
   align-items: center;
-  gap: 0.375rem;
+  gap: var(--space-1-5);
   border-radius: var(--radius-full, 999px);
   background: color-mix(in srgb, var(--surface-1) 88%, transparent);
-  padding: 0.125rem 0.5rem;
-  font-size: 0.6875rem;
+  padding: var(--space-0-5) var(--space-2);
+  font-size: var(--text-2xs);
   color: var(--text-secondary);
   backdrop-filter: blur(6px);
 }
